[FIX] Did a little adjustment and cleared some build files
