+++ hr_pms/data/data.xml
@@ -29,23 +29,23 @@
         </record>
         <!-- PMS YEAR -->
         <record model="pms.year" id="pms_year_2023_qrt1">
-            <field name="name">Jan-Mar 2023</field>
-            <field name="date_from">2023-01-01</field>
-            <field name="date_end">2023-03-30</field>
+            <field name="name">April 2023 [Quarterly]</field>
+            <field name="date_from">2023-04-01</field>
+            <field name="date_end">2023-04-01</field>
         </record>
         <record model="pms.year" id="pms_year_2023_qrt2">
-            <field name="name">April-June 2023</field>
-            <field name="date_from">2023-04-01</field>
-            <field name="date_end">2023-06-30</field>
+            <field name="name">July 2023 [Quarterly]</field>
+            <field name="date_from">2023-07-01</field>
+            <field name="date_end">2023-07-31</field>
         </record>
         <record model="pms.year" id="pms_year_2023_qrt3">
-            <field name="name">July-Sept 2023</field>
-            <field name="date_from">2023-07-01</field>
-            <field name="date_end">2023-09-30</field>
+            <field name="name">Oct 2023 [Quarterly]</field>
+            <field name="date_from">2023-10-01</field>
+            <field name="date_end">2023-10-31</field>
         </record>
         <record model="pms.year" id="pms_year_2023_qrt4">
-            <field name="name">Oct-Dec 2023</field>
-            <field name="date_from">2023-10-01</field>
+            <field name="name">Dec 2023 [End of year appraisal]</field>
+            <field name="date_from">2023-12-01</field>
             <field name="date_end">2023-12-30</field>
         </record>
